[fix] 충돌해결
