+++ app/(client)/(anon)/spots/SpotsPage.module.scss
