--- app/(client)/(anon)/spots/[spotId]/components/bookMark.module.scss
@@ -1,7 +1,0 @@
-.bookmarkButton {
-  background: none;
-  border: 0;
-  align-self: center;
-  width: 1.5rem;
-  height: 1.5rem;
-}
