--- app/(client)/(anon)/spots/[spotId]/docents/components/docentContent.module.scss
@@ -1,47 +1,0 @@
-.docentContainer {
-  border-bottom: 0.0625rem solid var(--grey-1-color);
-  padding: 1rem 0;
-
-  .docentWrapper {
-    display: flex;
-    gap: 0.625rem;
-    align-items: center;
-    margin-bottom: 0.625rem;
-
-    h3 {
-      font-size: 1rem;
-      font-weight: 600;
-    }
-  }
-
-  .contentWrapper {
-    display: flex;
-    flex-direction: column;
-
-    .clampedText,
-    .expandedText {
-      line-height: 1.25rem;
-      display: -webkit-box;
-      -webkit-line-clamp: 5;
-      line-clamp: 5;
-      -webkit-box-orient: vertical;
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
-
-    .expandedText {
-      display: block;
-      -webkit-line-clamp: unset;
-      line-clamp: unset;
-    }
-
-    .moreButton {
-      background: none;
-      border: none;
-      color: var(--grey-2-color);
-      cursor: pointer;
-      padding: 0;
-      align-self: flex-end;
-    }
-  }
-}
