--- app/(client)/(anon)/spots/[spotId]/docents/components/speakerAnimation.module.scss
@@ -1,51 +1,0 @@
-@keyframes waveAnimation {
-  0% {
-    opacity: 0;
-    transform: scale(0.8);
-  }
-
-  50% {
-    opacity: 1;
-    transform: scale(1.2);
-  }
-
-  100% {
-    opacity: 0;
-    transform: scale(0.8);
-  }
-}
-
-.container {
-  display: flex;
-  align-items: center;
-  font-size: 1.5rem;
-}
-
-.icon {
-  color: var(--primary-color);
-}
-
-
-.wave1 {
-  animation: waveAnimation 0.8s infinite ease-in-out;
-  animation-delay: 0.1s;
-  color: var(--primary-color);
-  font-size: 0.5rem;
-  margin-left: -0.4rem;
-
-}
-
-.wave2 {
-  animation: waveAnimation 0.8s infinite ease-in-out;
-  animation-delay: 0.2s;
-  color: var(--primary-color);
-  font-size: 0.8rem;
-}
-
-.wave3 {
-  animation: waveAnimation 0.8s infinite ease-in-out;
-  animation-delay: 0.3s;
-  color: var(--primary-color);
-  font-size: 1rem;
-  margin-bottom: 0.1rem;
-}
