--- app/(client)/(anon)/spots/[spotId]/docents/docentsPage.module.scss
@@ -1,32 +1,0 @@
-.noData {
-  padding: 1.25rem;
-  background-color: white;
-  flex-grow: 1;
-  gap: 0.625rem;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.docentsContainer {
-  padding: 0.625rem 1.25rem;
-  background-color: white;
-
-  .srOnly {
-    position: absolute;
-    width: 0.0625rem;
-    height: 0.0625rem;
-    padding: 0;
-    margin: -0.0625rem;
-    overflow: hidden;
-    clip-path: inset(50%);
-    border: 0;
-    clip: rect(0 0 0 0);
-  }
-
-  .docentWrapper {
-    background-color: white;
-  }
-}
