--- app/(client)/(anon)/spots/[spotId]/images/components/spotImage.module.scss
@@ -1,6 +1,0 @@
-.spotImage {
-    width: 10.625rem;
-    height: 10.625rem;
-    border-radius: 0.625rem;
-    object-fit: cover;
-}
