--- app/(client)/(anon)/spots/[spotId]/images/imagePage.module.scss
@@ -1,24 +1,0 @@
-.noData {
-  padding: 1.25rem;
-  background-color: white;
-  flex-grow: 1;
-  gap: 0.625rem;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  .link {
-    text-decoration:none; 
-    color: var(--primary-color);
-  }
-}
-
-.imageContainer {
-  padding: 1.25rem;
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  gap: 1.25rem;
-}
