--- app/(client)/(anon)/spots/[spotId]/info/components/ticketList.module.scss
@@ -1,59 +1,0 @@
-  .srOnly {
-    position: absolute;
-    width: 1px;
-    height: 1px;
-    padding: 0;
-    margin: -1px;
-    overflow: hidden;
-    clip-path: inset(50%);
-    border: 0;
-    clip: rect(0 0 0 0);
-  }
-
-  .ticketListContainer {
-    display: flex;
-    align-items: flex-start;
-
-    .icon {
-      flex-shrink: 0;
-      align-self: flex-start;
-      margin-right: 0.5rem;
-    }
-
-    .ticketListWrapper {
-      display: flex;
-      flex-direction: column;
-      width: 100%;
-      gap: 0.5rem;
-
-      li {
-        display: flex;
-        align-items: center;
-        width: 100%;
-
-        .label {
-          flex-shrink: 0;
-          margin-right: 0.5rem;
-        }
-
-        .line {
-          flex-grow: 1;
-          height: 1px;
-          background-color: var(--grey-2-color);
-          margin: 0 0.5rem;
-        }
-
-        .bold {
-          flex-shrink: 0;
-        }
-      }
-    }
-
-    button {
-      background-color: transparent;
-      border: none;
-      cursor: pointer;
-      margin-left: 0.5rem;
-      align-self: flex-start;
-    }
-  }
