--- app/(client)/(anon)/spots/[spotId]/info/infoPage.module.scss
@@ -1,74 +1,0 @@
-.infoContainer {
-  padding: 1.25rem;
-  background-color: white;
-
-  .srOnly {
-    position: absolute;
-    width: 0.0625rem;
-    height: 0.0625rem;
-    width: 0.0625rem;
-    height: 0.0625rem;
-    padding: 0;
-    margin: -0.0625rem;
-    overflow: hidden;
-    clip-path: inset(50%);
-    border: 0;
-    clip: rect(0 0 0 0);
-  }
-
-  .info {
-    padding: 1.25rem;
-    background-color: rgba(253, 187, 9, 0.2);
-    border-radius: 0.625rem;
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-  }
-
-  .subInfo {
-    border-top: 0.0625rem solid var(--grey-1-color);
-    font-size: 0.625rem;
-    color: var(--grey-2-color);
-    padding-top: 0.5rem;
-    li {
-      margin-bottom: 0.3rem;
-    }
-  }
-
-  .infoList {
-    display: flex;
-    flex-direction: column;
-    gap: 0.8rem;
-    margin-top: 1rem;
-
-    .copyWrapper {
-      display: flex;
-    }
-
-    .link {
-      color: var(--blue-flower-color);
-    }
-
-    p {
-      display: flex;
-      align-items: center;
-      gap: 0.5rem;
-
-      button {
-        background-color: transparent;
-        border: none;
-        cursor: pointer;
-      }
-
-      .subText {
-        font-size: 0.625rem;
-        color: var(--grey-2-color);
-      }
-
-
-      .primaryBold {
-        color: var(--primary-color);
-        font-weight: bold;
-      }
-    }
-  }
-}
