--- app/(client)/(anon)/spots/[spotId]/spotsLayout.module.scss
@@ -1,8 +1,0 @@
-.layout {
-  height: 100vh;
-  overflow-y: auto;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-}
-
