--- app/(client)/(anon)/spots/[spotId]/tips/components/tip.module.scss
@@ -1,74 +1,0 @@
-.tipContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  gap: 1rem;
-
-  .srOnly {
-    position: absolute;
-    width: 0.0625rem;
-    height: 0.0625rem;
-    padding: 0;
-    margin: -0.0625rem;
-    overflow: hidden;
-    clip-path: inset(50%);
-    border: 0;
-    clip: rect(0 0 0 0);
-  }
-
-  .profileWrapper {
-    display: flex;
-    gap: 0.625rem;
-    align-items: center;
-    justify-content: space-between;
-  }
-
-  .imageBorder {
-    border-radius: 50%;
-    border: 0.125rem solid var(--primary-color);
-    width: 2.25rem;
-    height: 2.25rem;
-  }
-}
-
-.costWrapper {
-  display: flex;
-  gap: 0.25rem;
-  font-weight: bold;
-  align-items: center;
-  justify-content: space-between;
-
-  .costBox {
-    display: flex;
-    gap: 0.25rem;
-
-    .costTitle {
-      color: var(--grey-2-color);
-    }
-
-    .cost {
-      color: var(--primary-color);
-    }
-  }
-
-  .createdAt {
-    color: var(--grey-2-color);
-    font-weight: 400;
-    font-size: 0.75rem;
-  }
-
-}
-
-.waitingTimeWrapper {
-  display: flex;
-  gap: 0.25rem;
-  align-items: center;
-
-  .waitingTimeTitle {
-    color: var(--grey-2-color);
-  }
-
-  .waitingTime {
-    color: var(--primary-color);
-  }
-}
