--- app/(client)/(anon)/spots/[spotId]/tips/tipsPage.module.scss
@@ -1,69 +1,0 @@
-.noData {
-  padding: 1.25rem;
-  background-color: white;
-  flex-grow: 1;
-  gap: 0.625rem;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.tipsContainer {
-  padding: 0.625rem 1.25rem;
-  background-color: white;
-
-  .srOnly {
-    position: absolute;
-    width: 0.0625rem;
-    height: 0.0625rem;
-    padding: 0;
-    margin: -0.0625rem;
-    overflow: hidden;
-    clip-path: inset(50%);
-    border: 0;
-    clip: rect(0 0 0 0);
-  }
-  .srOnly {
-    position: absolute;
-    width: 0.0625rem;
-    height: 0.0625rem;
-    padding: 0;
-    margin: -0.0625rem;
-    overflow: hidden;
-    clip-path: inset(50%);
-    border: 0;
-    clip: rect(0 0 0 0);
-  }
-
-  .sortContainer {
-    display: flex;
-    gap: 1.25rem;
-    justify-content: flex-end;
-
-    button {
-      background: none;
-      border: none;
-      cursor: pointer;
-      padding: 0;
-      font-size: 0.75rem;
-    }
-  }
-
-  .link {
-    color: var(--grey-2-color);
-    text-decoration: none;
-  }
-
-  .activeLink {
-    color: var(--primary-color);
-    text-decoration: none;
-  }
-
-  .tipContainer {
-    margin-bottom: 1.25rem;
-    border-bottom: 0.0625rem solid var(--grey-1-color);
-    padding-bottom: 1.25rem;
-  }
-}
