--- app/admin/components/sidebar/Sidebar.module.scss
@@ -1,29 +1,0 @@
-.sidebar {
-  width: 15rem;
-  padding: 2rem;
-  background-color: var(--primary-color);
-
-  .logo {
-    font-size: 2rem;
-    font-weight: bold;
-    text-align: center;
-  }
-
-  .navList {
-    list-style: none;
-    padding: 0;
-    margin-top: 2rem;
-
-    .navItem {
-      font-size: 1.2rem;
-      padding: 1rem;
-      cursor: pointer;
-    }
-  }
-}
-
-.active {
-  background: #000;
-  color: white;
-  border-radius: 0.5rem;
-}
