+++ app/admin/components/spotTable/SpotTable.module.scss
