+++ app/admin/components/tipTable/TipTable.module.scss
