--- app/admin/components/tiptable/Tiptable.module.scss
@@ -1,63 +1,0 @@
-.table {
-  width: 100%;
-  border-collapse: collapse;
-  background: white;
-  border-radius: 0.5rem;
-  overflow: hidden;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-
-  th {
-    background: var(--grey-1-color);
-    color: var(--grey-3-color);
-    font-weight: bold;
-    padding: 1rem;
-    text-align: left;
-    cursor: pointer;
-    transition: background 0.2s ease-in-out;
-
-    th.sorted {
-      background: var(--grey-2-color);
-      transform: scale(1.05);
-    }
-
-    tbody tr:hover {
-      background: var(--grey-1-color);
-      transition: background 0.2s ease-in-out;
-    }
-  }
-
-  td {
-    border-bottom: 1px solid var(--grey-1-color);
-    padding: 1rem;
-    text-align: left;
-    color: var(--grey-3-color);
-  }
-}
-
-.noData {
-  text-align: center;
-  padding: 1.5rem;
-  font-size: 1rem;
-  color: var(--grey-2-color);
-}
-
-.editButton,
-.deleteButton {
-  padding: 0.5rem 1rem;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 1rem;
-}
-
-.editButton {
-  background-color: var(--primary-color);
-  color: white;
-  margin-right: 0.5rem;
-}
-
-.deleteButton {
-  background-color: var(--red-1-color);
-  color: white;
-  transition: background 0.2s ease-in-out;
-}
